Support storing key location in redis.
We use a temporary (redis) set for keys that have been retrieved but
haven't been located so that we don't have to worry about authentication
when placing them. Only keys in this set can be placed in the actual
geospatial index which stores keys by location. Note that this code is
only the server side endpoints. We still need the client side UI to make
the request.