Use username + token pair to auth a user location change
A token is generated when the user authenticates with last.fm and it's
stored in redis set to expire an hour later. The token and username of
the user is passed to the client so it can make the request to
/set_user_location. Before, the session key was used as the primary way
to identify users, now usernames are used in the same way. That means
that we only have one location per user. Before each time they
authenticated the unique session key would make another pin, which
doesn't make any sense.